Split cluster to cluster_member and cluster_server
